moved Documents from management to sales
